tap_requirements.txt
requests==2.31.0
singer-python==5.13.0
